refactor: StaccatoShareService를 만들어 공유 기능 책임 분리
