+++ backend/src/main/java/com/staccato/staccato/service/StaccatoService.java
@@ -7,17 +7,12 @@ import com.staccato.staccato.service.dto.response.StaccatoDetailResponse;
 import com.staccato.staccato.service.dto.response.StaccatoIdResponse;
 import com.staccato.staccato.service.dto.response.StaccatoLocationResponse;
 import com.staccato.staccato.service.dto.response.StaccatoLocationResponses;
-import java.time.LocalDateTime;
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.staccato.comment.domain.Comment;
 import com.staccato.comment.repository.CommentRepository;
-import com.staccato.config.jwt.ShareTokenProvider;
-import com.staccato.config.jwt.dto.ShareTokenPayload;
 import com.staccato.config.log.annotation.Trace;
 import com.staccato.exception.ForbiddenException;
 import com.staccato.exception.StaccatoException;
@@ -28,9 +23,6 @@ import com.staccato.staccato.domain.StaccatoImage;
 import com.staccato.staccato.repository.StaccatoImageRepository;
 import com.staccato.staccato.repository.StaccatoRepository;
 import com.staccato.staccato.service.dto.request.FeelingRequest;
-import com.staccato.member.repository.MemberRepository;
-import com.staccato.staccato.service.dto.response.StaccatoShareLinkResponse;
-import com.staccato.staccato.service.dto.response.StaccatoSharedResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -44,10 +36,6 @@ public class StaccatoService {
     private final CategoryRepository categoryRepository;
     private final CommentRepository commentRepository;
     private final StaccatoImageRepository staccatoImageRepository;
-    private final ShareTokenProvider shareTokenProvider;
-    private final MemberRepository memberRepository;
-    @Value("${staccato.share.link-prefix}")
-    private String shareLinkPrefix;
 
     @Transactional
     public StaccatoIdResponse createStaccato(StaccatoRequest staccatoRequest, Member member) {
@@ -120,39 +108,9 @@ public class StaccatoService {
         staccato.changeFeeling(feeling);
     }
 
-    public StaccatoShareLinkResponse createStaccatoShareLink(Long staccatoId, Member member) {
-        Staccato staccato = getStaccatoById(staccatoId);
-        validateCategoryOwner(staccato.getCategory(), member);
-
-        ShareTokenPayload shareTokenPayload = new ShareTokenPayload(staccatoId, member.getId());
-        String token = shareTokenProvider.create(shareTokenPayload);
-        String shareLink = shareLinkPrefix + token;
-
-        return new StaccatoShareLinkResponse(staccatoId, shareLink, token);
-    }
-
-    public StaccatoSharedResponse readSharedStaccatoByToken(String token) {
-        shareTokenProvider.validateToken(token);
-        long staccatoId = shareTokenProvider.extractStaccatoId(token);
-        long memberId = shareTokenProvider.extractMemberId(token);
-        LocalDateTime expiredAt = shareTokenProvider.extractExpiredAt(token);
-
-        Staccato staccato = getStaccatoById(staccatoId);
-        Member member = getMemberById(memberId);
-        List<StaccatoImage> staccatoImages = staccatoImageRepository.findAllByStaccatoId(staccatoId);
-        List<Comment> comments = commentRepository.findAllByStaccatoId(staccatoId);
-
-        return new StaccatoSharedResponse(expiredAt, staccato, member, staccatoImages, comments);
-    }
-
     private Staccato getStaccatoById(long staccatoId) {
         return staccatoRepository.findById(staccatoId)
                 .orElseThrow(() -> new StaccatoException("요청하신 스타카토를 찾을 수 없어요."));
-    }
-
-    private Member getMemberById(long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new StaccatoException("요청하신 멤버를 찾을 수 없어요."));
     }
 
     private void validateCategoryOwner(Category category, Member member) {
